Make TUI and GPU features compile-time optional
Extract TUI code into src/tui.rs behind a "tui" feature flag.
Add "gpu" feature flag for blade-graphics GPU acceleration.
Default features include "tui" for backward compatibility.
