+++ Cargo.toml
@@ -14,17 +14,25 @@ path = "src/main.rs"
 name = "main"
 harness = false
 
+[features]
+default = []
+tui = ["dep:ratatui", "dep:crossterm"]
+gpu = ["dep:blade-graphics", "dep:blade-macros", "dep:bytemuck"]
+
 [dependencies]
 choir = { git = "https://github.com/kvark/choir" }
 crossbeam-channel = "0.5"
-crossterm = "0.28"
+crossterm = { version = "0.28", optional = true }
 log = "0.4"
 rand = "0.8"
 ron = "0.8"
 rustc-hash = "1"
 serde = { version = "1", features = ["serde_derive"]}
-ratatui = { version = "0.29", features = ["crossterm"] }
+ratatui = { version = "0.29", features = ["crossterm"], optional = true }
 gif = "0.13"
+blade-graphics = { git = "https://github.com/kvark/blade", optional = true }
+blade-macros = { git = "https://github.com/kvark/blade", optional = true }
+bytemuck = { version = "1", features = ["derive"], optional = true }
 
 [dev-dependencies]
 criterion = "0.5"
