+++ src/lib.rs
@@ -1,5 +1,7 @@
 pub mod analysis;
 pub mod grid;
+#[cfg(feature = "gpu")]
+pub mod gpu;
 pub mod lab;
 pub mod narrative;
 pub mod render;
